% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAPITR_TestData_Pathways}
\alias{MAPITR_TestData_Pathways}
\title{MAPITR TestData Pathways}
\format{
A list with 5 entries, vectors of 75 variables each:

\describe{
  \item{Pathway}{Simulated
     pathway number}
}
}
\source{
Manually created
}
\description{
A simulated set of pathways for use in unit
tests.
}
\keyword{data}
