% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAPITR_TestData_PCs}
\alias{MAPITR_TestData_PCs}
\title{MAPITR TestData PCs}
\format{
A data frame with 500 rows (individuals) and 10 
variables (PCs)
}
\source{
Manually created
}
\description{
Top 10 principal components from the simulated
genotypes for for use in unit tests
}
\keyword{data}
