% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAPITR_TestData_Phenotype}
\alias{MAPITR_TestData_Phenotype}
\title{MAPITR TestData Phenotype}
\format{
A vector with 750 rows and 1 variables:

\describe{
  \item{Phenotype}{Simulated
     phenotype values}
}
}
\source{
Manually created
}
\description{
A simulated set of phenotypes for use in unit
tests.
}
\keyword{data}
