% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MAPITRBase}
\alias{MAPITRBase}
\title{Main MAPITR cpp code}
\usage{
MAPITRBase(X, Y, regions, cores = 1L)
}
\arguments{
\item{X}{a p x n genotype matrix}

\item{Y}{a 1 x n phenotype vector}

\item{regions}{a list of q pathways}

\item{cores}{an integer for the number of cores}
}
\description{
Main MAPITR cpp code
}
