% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAPITR_SimData_Pathways}
\alias{MAPITR_SimData_Pathways}
\title{MAPITR SimData Pathways}
\format{
A data frame with 5 rows (pathways) and 2 variables 
(pathway name, comma-separated list of pathway column indices)
}
\source{
Manually created
}
\description{
A simulated set of pathways for use in 
vignettes
}
\keyword{data}
