% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAPITR_TestData_Genotypes}
\alias{MAPITR_TestData_Genotypes}
\title{MAPITR TestData Genotypes}
\format{
A data frame with 500 rows (individuals) and 750 
variables (SNPs)
}
\source{
Manually created
}
\description{
A simulated dataset of genotypes for use in unit
tests
}
\keyword{data}
