% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAPITR_TestData_Phenotype}
\alias{MAPITR_TestData_Phenotype}
\title{MAPITR TestData Phenotype}
\format{
A data frame with 500 rows (individuals) and 1 
variables (phenotype)
}
\source{
Manually created
}
\description{
A simulated set of phenotypes for use in unit
tests
}
\keyword{data}
